requirements.txt
PySide6>=6.6.0
requests>=2.31.0
psutil>=5.9.6
python-bitcoinrpc>=1.0
scrypt>=0.8.24
base58>=2.1.1
pyinstaller>=6.0.0
matplotlib>=3.5.0